fix(jest-config): throw correct error for missing preset modules
When loading a preset module that cannot be resolved (e.g. is not npm
installed), the following error is shown to users:

    ● Validation Error:

      Preset foobar is invalid:

      The "id" argument must be of type string. Received null

Instead, the following error should be shown

    ● Validation Error:

      Preset foobar not found.

This PR changes the error message by handling when
Resolver.findNodeModule returns null, indicating that the module cannot
be successfully imported.